Add the ability to render ordered list
We refactored our ordered list handling, as it was handling a lot of state in
order to determine when a list ends, and the enumerable module can handle that
for us.

We're doing a lot of leg work to determine if there is a numeric prefix,
because the project spec says we aren't allowed to use regular expressions.